move sizes of binaries into the buffer.
this is actually worse when the binaries (or atoms) are all the same
size, as schema compression was getting rid of them. however, when
the binaries in question are different lengths this turns out better as
schema compression can compress the type codes regardless of length.

this is typically what is seen in the real world, so is probably the
better compromise.